fix(37-08): resolve all TODOs and placeholders per TigerStyle
- Router: WinterCGHandler now returns 503 instead of fake Phase 3 text
- Module loader: VFS properly passed through execution context
- Crypto ECDH: Implemented using p256/p384 ecdh features
- V8 version: Uses actual v8::V8::get_version() instead of hardcoded string
- OOM monitor: Fixed hostname logging with MemoryLimiter.hostname() getter
- VFS capture: Implemented recursive file capture using list_dir/read
- Metrics: Removed empty PrometheusMetricFamily placeholder struct
- CLI errors: Removed commented-out helper constructors
- Assertions: Documented design-time enforcement macro pattern
- Tests: Updated http/client test comments, removed duplicate tests
- Various: Documented intentional cold sliver and legacy snapshot design

+++ src/main.rs
@@ -296,7 +296,7 @@ async fn run_from_sliver(
     // Create SliverWorkerPool with temp entrypoint for VFS-extracted JS
     let worker_pool = Arc::new(nano::worker::pool::SliverWorkerPool::with_temp_entrypoint(
         hostname.clone(),
-        workers,
+        workers as u32,
         0, // No memory limit for now
         unpacked,
         Some(temp_entrypoint.clone()),
